check chipId switch to platform
Match-id-854607350a4492bf0e370755a0a62831e8d54975

+++ collector/dvvp/common/config/config_manager.cpp
@@ -60,7 +60,7 @@ int ConfigManager::Init()
             return PROFILING_FAILED;
         }
     } else {
-        chipId = ((static_cast<uint32_t>(versionInfo) >> 8) & 0xff); // 8:shift 8 bits, get the low 8 bits(0xff)
+        chipId = ((static_cast<uint64_t>(versionInfo) >> 8) & 0xff); // 8:shift 8 bits, get the low 8 bits(0xff)
         if (chipId >= PlatformType::END_TYPE) {
             MSPROF_LOGE("halGetDeviceInfo get chip invalid, versionInfo:0x%llx, chipId:%u", versionInfo, chipId);
         }
